% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedmat.r
\name{speedmat.jsdist}
\alias{speedmat.jsdist}
\title{SpEED matrix by multiplying Jensen-Shannon divergence and geodesic distance (in km)}
\usage{
speedmat.jsdist(data, formula, outcome = "outcome", treatment = "treatment")
}
\arguments{
\item{data}{sf object. Should include outcome, treatment, and coordinates}

\item{formula}{formula. in \code{y ~ x} form.}

\item{outcome}{character(1). Outcome variable name. default is \code{'outcome'}}

\item{treatment}{character(1). Treatment variable name. default is \code{'treatment'}}
}
\description{
Returns a SpEED matrix, which is the product of Jensen-Shannon divergence and geographic distance matrix
}
\author{
Insang Song (sigmafelix@hotmail.com)
}
