% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distJSD}
\alias{distJSD}
\title{Jensen-Shannon divergence}
\usage{
distJSD(inMatrix)
distJSD2(inMatrixTr, inMatrixCo)
}
\arguments{
\item{inMatrix}{a positive-valued numeric matrix.}
\item{inMatrixTr}{a positive-valued numeric matrix including covariates of treated units.}
\item{inMatrixCo}{a positive-valued numeric matrix including covariates of controlled units.}

}
\description{
Computes Jensen-Shannon divergence, which is based on Kullback-Leibler divergence. \eqn{JSD(A \Vert B)=\frac{1}{2} \left( KL(A \Vert M)+KL(B \Vert M) \right)}
}
